pyproject.toml
[project]
name = "ai-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "elevenlabs>=2.11.0",
    "openai>=1.101.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "streamlit>=1.48.1",
]
